streamlit_app/requirements.txt
streamlit>=1.40.0
sqlmodel>=0.0.14
pydantic>=2.0
google-generativeai
fpdf2
plotly
python-dotenv
gspread
google-auth
streamlit-agraph
